Transformando parser em tokenizer
